feat: renamed package
- Renamed cmfive package to gen.
