+++ main.go
@@ -2,10 +2,10 @@ package main
 
 import (
 	"github.com/strongishllama/cmfive-cli/pkg/cli"
-	"github.com/strongishllama/cmfive-cli/pkg/cmfive"
+	"github.com/strongishllama/cmfive-cli/pkg/gen"
 )
 
 func main() {
-	cmfive.TemplatesDir = "cmfive/"
+	gen.TemplatesDir = "gen/"
 	cli.Execute()
 }
