--- pkg/cmfive/migration_test.go
@@ -1,17 +1,0 @@
-package cmfive_test
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/require"
-
-	"github.com/strongishllama/cmfive-cli/pkg/cmfive"
-)
-
-func TestNewMigration(t *testing.T) {
-	setup(t)
-	defer teardown(t)
-
-	require.NoError(t, cmfive.NewModule("payroll"))
-	require.NoError(t, cmfive.NewMigration("payroll", "InitialMigration"))
-}
