+++ pkg/gen/module_test.go
@@ -1,16 +1,16 @@
-package cmfive_test
+package gen_test
 
 import (
 	"testing"
 
 	"github.com/stretchr/testify/require"
 
-	"github.com/strongishllama/cmfive-cli/pkg/cmfive"
+	"github.com/strongishllama/cmfive-cli/pkg/gen"
 )
 
 func TestNewModule(t *testing.T) {
 	setup(t)
 	defer teardown(t)
 
-	require.NoError(t, cmfive.NewModule("payroll"))
+	require.NoError(t, gen.NewModule("payroll"))
 }
