+++ pkg/gen/templates/README.tmpl
