+++ pkg/gen/templates/action.tmpl
