+++ pkg/gen/templates/config.tmpl
