+++ pkg/gen/templates/hooks.tmpl
