+++ pkg/gen/templates/roles.tmpl
