+++ pkg/gen/templates/service.tmpl
